[CI] Update MacOS Deployment target to fix failing builds (#140)
The OSX+Win nightly builds are broken for several month now due to an incompatibility with older MacOS SDK versions in TVM.

```
 [ 15%] Building CXX object CMakeFiles/tvm_runtime_objs.dir/src/runtime/debug.cc.o
/Users/runner/work/tlcpack/tlcpack/tvm/src/arith/conjunctive_normal_form.cc:363:50: error: 'value' is unavailable: introduced in macOS 10.13
          auto& key_i = i_chunk[i_distinct_index.value()];
                                                 ^
/Applications/Xcode_13.2.1.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX12.1.sdk/usr/include/c++/v1/optional:954:27: note: 'value' has been explicitly marked unavailable here
    constexpr value_type& value() &
                          ^
```

I "fixed" this by passing the `MACOSX_DEPLOYMENT_TARGET` also the script which build the TVM lib. While this probably drops a little bit of backwards compatibility on MacOS, IMHO it is still better to have a successful build instead of a failing one.

A TVM-side bugfix could be an alternative solution for this issue.