Fix Python version check when Conda not installed
When running tlcpack on build environments that do not have Conda
installed, the script is failing due to the method used to locate the
Python version.

5382fa075932676938728a241e21bfcb8e714ea1 changed the way in which the
Python version is located, now hardcoding a path that only exists in
environments where Conda has been installed.

I have added a check to fall back to the previous method of locating the
Python version on environments where Conda is not installed.