Add logic to skip the Git checkout
In some circumstances we would like to skip the checkout stage, in order
to run the package sync on an existing checkout.

In this case it is not necessary and in some cases, undesirable to run
the checkout again so this patch adds logic to bypass it when required.

Unless the --skip-checkout option is passed, existing checkout logic
remains unchanged by this patch.