Fix Docker Cleanup requires build agent
The cleanup step was placed in the post execution steps. Because a shell
command is being called to cleanup the Docker images, it requires the
use of a build agent. As a global agent was not configured, these shell
commands failed to execute. The execution of this cleanup has been moved
to the end of the stage that builds the wheels.