Add switch to toggle specific build types
It may not always be desired to build both CPU and AArch64 packages
every time this pipeline is run. A boolean switch has been added for
both. When the switch for an architecture type is left unchecked, it
will skip build steps for that architecture and exclude it from the
package.