[CI] Add new PyPI packaging Jenkins pipeline (#168)
* Add new PyPI packaging Jenkins pipeline

This patch adds a new Jenkins pipeline for packaging. This is a
migration of the current Github actions workflow of uploading a Tensor
learning compiler binary distribution (tlcpack) to PyPI.

This new packaging pipeline implements a parallel step that creates and
uploads AArch64 packages, which is something not present in the Github
actions job it is intended to replace.

* Fix Docker Cleanup requires build agent

The cleanup step was placed in the post execution steps. Because a shell
command is being called to cleanup the Docker images, it requires the
use of a build agent. As a global agent was not configured, these shell
commands failed to execute. The execution of this cleanup has been moved
to the end of the stage that builds the wheels.

* Fix sudo - a terminal required to read password error

Removed the instruction to download the latest version of pip from the
pipeline, as it was failing with a sudo permissions error. This was not
required in the tlcpack Jenkins environment, only in my local testing
environment.

* Add switch to toggle specific build types

It may not always be desired to build both CPU and AArch64 packages
every time this pipeline is run. A boolean switch has been added for
both. When the switch for an architecture type is left unchecked, it
will skip build steps for that architecture and exclude it from the
package.

* Refactor pipeline to use S3 bucket storage

Made changes to the pipeline to replace all instances of Jenkins
stashing with an upload to and download from S3 cloud storage instead.

* Cut CPU packaging support

Removed all support for generating CPU packages, leaving only AArch64
tasks in the pipeline.