Cut daily Docker image validate trigger (#176)
Since work was completed to split up Jenkins in to platform-specific
jobs (PR: https://github.com/apache/tvm/pull/13316), the pipeline for
validating the docker images with tests has been failing.

This validation pipeline is no longer needed and can be disabled,
however in doing so, it causes a failure of the
`daily-docker-image-rebuild` pipeline, which triggers the validation
pipeline to be run if the build is successful. This is because the
validation pipeline relied on the file `JenkinsFile`, which was removed
as part of the changes mentioned above.

In order to resolve the build errors for the daily docker rebuild, it is
necessary to remove the trigger for the validation pipeline.