[CI] Modify PyPI packaging Jenkins pipeline (#179)
* Modify PyPI packaging Jenkins pipeline

Changed the pipeline to enable the user to specify which Git ref to use
for the checkout and package build. Removed additional checkout logic
causing issues building the correct package.

Parameterised other variables which were previously only able to be
modified within the script. Configured these with sensible default
values that can be modified via the Jenkins user interface.

Minor changes to formatting.

* Modify PyPI packaging Jenkins pipeline

Changed the AArch64 Docker hash to the latest available.