Pin Cython fix for compilation
Cython 3.0.0 was recently released, but it is incompatible with the
current .pxi definitions in python/tvm/_ffi/_cpython.

Cython has been pinned for TVM in PR: 15353 until a working fix is
created. This patch replicates many of the changes for TLCPack
installation, as we are seeing the same Cython errors when trying to use
TLCPack Ubuntu Docker images.